+++ envi/symstore/resolver.py
@@ -20,7 +20,7 @@ class Symbol:
     def __eq__(self, other):
         if not isinstance(other, Symbol):
             return False
-        return long(self) == long(other)
+        return int(self) == int(other)
 
     def __coerce__(self, value):
         t = type(value)
@@ -32,13 +32,13 @@ class Symbol:
             return (t(self.value), value)
 
         if isinstance( value, Symbol ):
-            return ( long(self.value), long(value.value) )
+            return ( int(self.value), int(value.value) )
 
     def __hash__(self):
-        return hash(long(self))
+        return hash(int(self))
 
     def __long__(self):
-        return long(self.value)
+        return int(self.value)
 
     def __int__(self):
         return int(self.value)
@@ -102,7 +102,7 @@ class SymbolResolver:
         """
         Delete a symbol from the resolver's namespace
         """
-        symval = long(sym)
+        symval = int(sym)
         self.symaddrs.pop(symval, None)
 
         # bbase = symval & self.bucketmask
